move manifest stuff
